moved metric card utilities to metric card component.
