+++ apps/festival/festival/src/app/dashboard/analytics/buyer/buyer-analytics.component.ts
@@ -6,10 +6,9 @@ import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { AnalyticsService } from '@blockframes/analytics/service';
 import { aggregate, counter, countedToAnalyticData } from '@blockframes/analytics/utils';
-import { MetricCard } from '@blockframes/analytics/components/metric-card-list/metric-card-list.component';
+import { MetricCard, events, toCards } from '@blockframes/analytics/components/metric-card-list/metric-card-list.component';
 import {
   AggregatedAnalytic,
-  EventName,
   isScreening,
   Invitation,
   Analytics,
@@ -19,7 +18,6 @@ import {
 } from '@blockframes/model';
 import { fromOrgAndAccepted, MovieService } from '@blockframes/movie/service';
 import { OrganizationService } from '@blockframes/organization/service';
-import { IconSvg } from '@blockframes/ui/icon.service';
 import { NavigationService } from '@blockframes/ui/navigation.service';
 import { UserService } from '@blockframes/user/service';
 import { App, sum, toLabel } from '@blockframes/model';
@@ -41,49 +39,6 @@ import { EventService } from '@blockframes/event/service';
 
 interface MovieWithAnalytics extends Movie { analytics: Analytics<'title'>[]; };
 
-
-export interface VanityMetricEvent {
-  name: EventName;
-  title: string;
-  icon: IconSvg;
-};
-
-const events: VanityMetricEvent[] = [
-  {
-    name: 'pageView',
-    title: 'Views',
-    icon: 'visibility'
-  },
-  {
-    name: 'promoReelOpened',
-    title: 'Promoreel Opened',
-    icon: 'star_fill'
-  },
-  {
-    name: 'addedToWishlist',
-    title: 'Adds to Wishlist',
-    icon: 'favorite'
-  },
-  {
-    name: 'screeningRequested',
-    title: 'Screening Requested',
-    icon: 'ask_screening_2'
-  },
-  {
-    name: 'askingPriceRequested',
-    title: 'Asking Price Requested',
-    icon: 'local_offer'
-  }
-];
-
-function toCards(aggregated: AggregatedAnalytic): MetricCard[] {
-  return events.map(event => ({
-    title: event.title,
-    value: aggregated[event.name],
-    icon: event.icon,
-    selected: false
-  }));
-}
 
 function filterAnalytics(title: string, analytics: AggregatedAnalytic[]) {
   const name = events.find(event => event.title === title)?.name;
